+++ src/Position.cpp
@@ -1,6 +1,6 @@
 #include "Position.h"
 
-Position::Position(size_t pos_x, size_t pos_y):
+Position::Position(int32_t pos_x, int32_t pos_y):
     x(pos_x),
     y(pos_y),
     x_d(0),
