+++ src/Sprites.h
@@ -10,8 +10,8 @@ class Sprite: public Size
 {
     public:
         const uint8_t* data;
-        Sprite(size_t, size_t, const uint8_t*);
-        size_t get_length(void);
+        Sprite(int32_t, int32_t, const uint8_t*);
+        int32_t get_length(void);
 };
 
 class Sprites
